pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sprintlens"
version = "0.1.0"
description = "프로그램팀 스프린트 일정 진행상황 리포트 웹 서비스"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }

dependencies = [
    "flask>=3.1.0",
    "atlassian-python-api>=3.41.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
]

[tool.setuptools.packages.find]
include = ["sprintlens*"]

# === Ruff Configuration ===
[tool.ruff]
target-version = "py313"
line-length = 88
include = ["app.py", "sprintlens/**/*.py", "tests/**/*.py"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade (modern Python syntax)
    "B",   # flake8-bugbear (potential bugs)
    "SIM", # flake8-simplify (code simplification)
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501", # line-length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["sprintlens"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# === pytest Configuration ===
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=sprintlens",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]
markers = [
    "slow: 시간이 오래 걸리는 테스트",
    "integration: 외부 API 연동 테스트",
]

# === Coverage Configuration ===
[tool.coverage.run]
source = ["sprintlens"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
